+++ src/Sql/Sql/ChangeLog.md
@@ -27,6 +27,8 @@
     - Enable-AzSqlInstanceDatabaseSensitivityRecommendation.
     - Disable-AzSqlDatabaseSensitivityRecommendation.
     - Disable-AzSqlInstanceDatabaseSensitivityRecommendation.
+* Remove client-side validation of 'RetentionDays' parameter from cmdlet SetAzureRmSqlDatabaseBackupShortTermRetentionPolicy
+* Auditing to a storage account in Vnet, fixing a bug when creating a Storage Blob Data Contributor role.
 
 ## Version 2.6.0
 * Added cmdlets `Get-AzSqlInstanceOperation` and `Stop-AzSqlInstanceOperation`